chore(select-text): place default values in a separate file for controller
