+++ packages/select-text/controller/src/defaults.js
@@ -1,0 +1,26 @@
+const tokens = () => [
+  {
+    text: 'Rachel cut out 8 stars in 6 minutes.',
+    start: 0,
+    end: 36,
+  },
+  {
+    text: 'Lovelle cut out 6 stars in 4 minutes.',
+    start: 37,
+    end: 74,
+  },
+  {
+    text: 'Lovelle and Rachel cut the same number of stars in 6 minutes.',
+    start: 117,
+    end: 177
+  }
+];
+
+export default {
+  highlightChoices: true,
+  maxSelections: 2,
+  prompt: 'What sentences contain the character 6 in them?',
+  text:
+    'Rachel cut out 8 stars in 6 minutes. Lovelle cut out 6 stars in 4 minutes. Rachel cut out 4 more stars than Lovelle. Lovelle and Rachel cut the same number of stars in 6 minutes.',
+  tokens: tokens(),
+};
